HealthPopup, SceneLoaded
Added to flow that the initalization of a flow happens only after sceneLoaded

HealthPopup now a feature. Pass HealthPopUpManager your transform and the change in hp and it will create a text over head that smoothdamps to value. If you add more dmg while it's active it will add it to it. Auto manages in manager, no need to call remove or etc